use schema files to define meta field of SolrObject and move parentList to components
