+++ meta/helper/solrObjects.ts
@@ -1,39 +1,15 @@
 import { SolrObject } from "../interface/SolrObject";
 import { SolrParent } from "../interface/SolrParent";
-import { findFirstSentence} from "./util";
+import aardvark_json from "../../src/pages/search/_metadata/aardvark_schema.json";
+import sdoh_json from "../../src/pages/search/_metadata/sdohplace_schema.json";
+import { AardvarkSdohSchemaMatch, findFirstSentence } from "./util";
 
 /**
- * Match between Aardvark and our Solr metadata,
- * except the required fields, which are defined in the SolrObject interface
+ *
+ * @param rawSolrObject
+ * @param matchingRule
+ * @returns
  */
-export const rawSolrMetdataSwitch = {
-	dct_language_sm: "language",
-	gbl_resourceType_sm: "resource_type",
-	dct_subject_sm: "subject",
-	dcat_theme_sm: "theme",
-	dct_issued_s: "date_issued",
-	gbl_indexYear_im: "index_year",
-	dct_isVersionOf_sm: "is_version_of",
-	gbl_suppressed_b: "suppressed",
-	_version_: "version",
-	dct_format_s: "format",
-	dct_creator_sm: "creator",
-	dct_publisher_sm: "publisher",
-	dct_references_s: "references",
-	dct_description_sm: "description",
-	dct_temporal_sm: "temporal_coverage",
-	gbl_displayNote_sm: "display_note",
-	dct_rights_sm: "rights",
-	// Following are in our object but not in Aardvark
-	sdoh_spatial_resolution_sm: "spatial_resolution", // spatial coverage?
-	timestamp: "timestamp",
-	score: "score",
-	schema_provider_s: "schema_provider",
-	sdoh_methods_variables_sm: "sdoh_methods_variables",
-	sdoh_data_variables_sm: "sdoh_data_variables",
-	locn_geometry: "location_geometry",
-};
-
 const initSolrObject = (rawSolrObject: any): SolrObject => {
 	let result = {} as SolrObject;
 	result.id = rawSolrObject.id;
@@ -54,7 +30,9 @@ const initSolrObject = (rawSolrObject: any): SolrObject => {
 			key !== "gbl_mdAccessRights_s" &&
 			key !== "gbl_resourceClass_s"
 		) {
-			result.meta[rawSolrMetdataSwitch[key]] = rawSolrObject[key];
+			result.meta[
+				AardvarkSdohSchemaMatch(key, aardvark_json, sdoh_json)
+			] = rawSolrObject[key];
 		}
 	});
 	return result;
